+++ src/Types/ContractIdentifier.cs
@@ -1,0 +1,7 @@
+namespace Concordium.Sdk.Types;
+
+/// <summary>
+/// Contains the identification name of the contract.
+/// </summary>
+/// <param name="ContractName">Name of contract</param>
+public sealed record ContractIdentifier(string ContractName);