+++ src/Types/EntryPoint.cs
@@ -1,0 +1,7 @@
+namespace Concordium.Sdk.Types;
+
+/// <summary>
+/// Entry point on a smart contract.
+/// </summary>
+/// <param name="Name">Entry point name</param>
+public sealed record EntryPoint(string Name);